rename: sign-up-tab.d.ts -> sign-in.d.ts, sign-up-tab.tsx -> sign-in.tsx
