+++ src/@types/sign-in.d.ts
@@ -3,8 +3,8 @@ export type TTab = "signIn" | "signUp"
 export type TRoute = "/my" | "/rolling-paper/write"
 
 export type TStore = {
-  signUpTab: TTab
-  setSignUpTab: (tab: TTab) => void
+  tab: TTab
+  setTab: (tab: TTab) => void
   redirectTo: TRoute
   setRedirectTo: (to: TRoute) => void
 }
