+++ src/store/sign-in.tsx
@@ -2,23 +2,21 @@
 
 import {useState, createContext, type ReactNode} from "react"
 import ROUTE from "@/constants/route"
-import type {TStore, TTab, TRoute} from "@/@types/sign-up-tab"
+import type {TStore, TTab, TRoute} from "@/@types/sign-in"
 
 const Store = createContext<TStore>({
-  signUpTab: "signIn",
-  setSignUpTab: function () {},
+  tab: "signIn",
+  setTab: function () {},
   redirectTo: ROUTE.MY_PAGE,
   setRedirectTo: function () {},
 })
 
 export const Provider = ({children}: {children: ReactNode}) => {
-  const [signUpTab, setSignUpTab] = useState<TTab>("signIn")
+  const [tab, setTab] = useState<TTab>("signIn")
   const [redirectTo, setRedirectTo] = useState<TRoute>(ROUTE.MY_PAGE)
 
   return (
-    <Store.Provider
-      value={{signUpTab, setSignUpTab, redirectTo, setRedirectTo}}
-    >
+    <Store.Provider value={{tab, setTab, redirectTo, setRedirectTo}}>
       {children}
     </Store.Provider>
   )
